+++ App/src/models/dossier.ts
@@ -1,0 +1,15 @@
+export interface DossierDTO {
+    id: string;
+    initiatorId: string;
+    title: string | null;
+    description: string | null;
+    published: boolean;
+}
+
+export interface DossierDTOResponse {
+    data: DossierDTO;
+}
+
+export interface GetMyDossiersResponse {
+    data: DossierDTO[];
+}
