+++ App/src/pages/Register.tsx
@@ -16,7 +16,7 @@ import {
 } from "@chakra-ui/react";
 import logo from "../assets/logo.png";
 import { useForm } from "react-hook-form";
-import { User } from "../services/user";
+import { User } from "../models/user";
 import { useNavigate } from "react-router-dom";
 import { useState } from "react";
 import { AuthenticationResponse, LoginProps, RegisterDTO, RegisterUser, decodeTokenToUser } from "../services/auth";
